backend/pyproject.toml
[project]
name = "tellmenow-backend"
version = "0.1.0"
description = "TellMeNow: Skill-based Q&A Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "aiosqlite>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tellmenow"]
